Multiple tokens (#7)
* router contract introduced


* unit tests and some improvements


* namings


* FeeDistributor contract typo


* wrappedId added in bridge, require msgs, typos, FeeDistributor marked as abstract


* public mint added without tx reimbursement, tests added, refactoring


* require -> modifier
