+++ contracts/Governance.sol
@@ -1,7 +1,6 @@
 pragma solidity ^0.6.0;
 
 import "@openzeppelin/contracts/access/Ownable.sol";
-import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/utils/EnumerableSet.sol";
 
 /**
@@ -9,24 +8,11 @@ import "@openzeppelin/contracts/utils/EnumerableSet.sol";
  *  @title Governance contract, providing governance/members functionality
  */
 abstract contract Governance is Ownable {
-    using SafeMath for uint256;
+    // using SafeMath for uint256;
     using EnumerableSet for EnumerableSet.AddressSet;
 
     /// @notice Iterable set of members
     EnumerableSet.AddressSet private membersSet;
-
-    /// @notice Total checkpoints so far
-    uint256 public totalCheckpoints;
-
-    /// @notice Total fees that could be claimed
-    uint256 public totalClaimableFees;
-
-    /// @notice Total fees accrued per each checkpoint
-    mapping(uint256 => uint256) public checkpointServiceFeesAccrued;
-
-    /// @notice Mapping of members and the fees that they are eligble to claim.
-    /// Does not include the fees of each member from the current checkpoint.
-    mapping(address => uint256) public claimableFees;
 
     /// @notice An event emitted once member is updated
     event MemberUpdated(address member, bool status);
@@ -42,9 +28,7 @@ abstract contract Governance is Ownable {
      * @param account The account to be modified
      * @param isMember Whether the account will be set as member or not
      */
-    function updateMember(address account, bool isMember) public onlyOwner {
-        createNewCheckpoint();
-
+    function _updateMember(address account, bool isMember) internal {
         if (isMember) {
             require(
                 membersSet.add(account),
@@ -72,52 +56,5 @@ abstract contract Governance is Ownable {
     /// @notice Returns the address of a member at a given index
     function memberAt(uint256 index) public view returns (address) {
         return membersSet.at(index);
-    }
-
-    /// @notice Creates a new checkpoint, distributing the accrued fees
-    /// of the previous checkpoint to all members
-    function createNewCheckpoint() internal {
-        uint256 mCount = membersCount();
-        if (mCount == 0) {
-            return;
-        }
-
-        uint256 feesAccrued = checkpointServiceFeesAccrued[totalCheckpoints];
-        uint256 feePerMember = feesAccrued.div(mCount);
-        if (feePerMember == 0) {
-            return;
-        }
-
-        uint256 feesTotal = feePerMember.mul(mCount);
-        uint256 feesLeft = feesAccrued.sub(feesTotal); // fees left due to integer division
-        totalCheckpoints++;
-        checkpointServiceFeesAccrued[totalCheckpoints] = feesLeft;
-
-        for (uint256 i = 0; i < mCount; i++) {
-            address currentMember = memberAt(i);
-            claimableFees[currentMember] = claimableFees[currentMember].add(
-                feePerMember
-            );
-        }
-    }
-
-    /**
-     * @notice Gets the claimable fee of a member, including the fee from the current checkpoint
-     * @param _address the target address
-     */
-    function claimableFeesFor(address _address) public view returns (uint256) {
-        if (!isMember(_address)) {
-            return claimableFees[_address];
-        }
-
-        uint256 currentCheckpointFeePerMember = 0;
-        if (membersCount() > 0) {
-            currentCheckpointFeePerMember = checkpointServiceFeesAccrued[
-                totalCheckpoints
-            ]
-                .div(membersCount());
-        }
-
-        return claimableFees[_address].add(currentCheckpointFeePerMember);
     }
 }
